+++ packages/core/src/interpolation/interpolateUrl.ts
@@ -4,9 +4,9 @@ function slugify(text: string) {
   return text.trim().replaceAll(' ', '-').replaceAll('/', '-').toLowerCase()
 }
 
-function interpolateUrl(url: string, values: Record<string, string>) {
-  Object.keys(values).forEach(key => {
-    const value = values[key];
+function interpolateUrl(url: string, context: Record<string, string>) {
+  Object.keys(context).forEach(key => {
+    const value = context[key];
     const slugifiedValue = slugify(value);
 
     url = interpolateValue(url, key, slugifiedValue);