+++ packages/core/src/resolve/resolveToString.ts
@@ -2,14 +2,14 @@ import { resolve, type ResolveCallback, type ResolveFunction } from './'
 
 const resolveToString: ResolveFunction<string> = (
   text,
-  values = {},
+  context = {},
   options,
 ) => {
   const cb: ResolveCallback<string> = (c) => {
     return c.prev + c.part
   }
 
-  return resolve(text, values, cb, '', options)
+  return resolve(text, context, cb, '', options)
 }
 
 export default resolveToString