+++ packages/core/src/translation/collect.ts
@@ -1,9 +1,9 @@
-import { Values } from '../interpolation'
+import { Context } from '../interpolation'
 import { RegisteredMessages } from '../localization'
 import { collection } from './'
 
-function collect(key: keyof RegisteredMessages, values?: Values) {
-  collection[key] = values || {}
+function collect(key: keyof RegisteredMessages, context?: Context) {
+  collection[key] = context || {}
 }
 
 export default collect