+++ packages/react/src/hooks/useI18n.ts
@@ -7,8 +7,8 @@ import useLocale from './useLocale'
 function useI18n() {
   const [locale, setLocale] = useLocale()
 
-  const t: typeof tLocale = (text, values, options) => {
-    return tLocale(text, values, options || { locale })
+  const t: typeof tLocale = (text, context, options) => {
+    return tLocale(text, context, options || { locale })
   }
 
   const dn: typeof dnLocale = (code, intlOptions, options) => {
@@ -19,8 +19,8 @@ function useI18n() {
     return vLocale(value, options || { locale })
   }
 
-  const u: typeof uLocale = (url, values, options) => {
-    return uLocale(url, values, options || { locale })
+  const u: typeof uLocale = (url, context, options) => {
+    return uLocale(url, context, options || { locale })
   }
 
   return { locale, setLocale, t, dn, v, u }