+++ packages/vue/src/components/T.vue
@@ -1,39 +1,41 @@
 <script setup lang="ts"
-  generic="T extends Register, K extends keyof T['messages'], V extends InferValues<K | E>, E extends T['messages'][K][keyof T['messages'][K]]">
-  import type { FormatContext, InferValues, Register, Values } from '@psitta/core'
-  import { collect, formatToSegments, prepareFormat } from '@psitta/core'
+  generic="T extends Register, K extends keyof T['messages'], V extends InferContext<K | E>, E extends T['messages'][K][keyof T['messages'][K]]">
+  import type { ResolveContext, InferContext, Register, Context } from '@psitta/core'
+  import { collect, localizeMessage, resolveToSegments } from '@psitta/core'
   import { computed, onMounted, useSlots } from 'vue'
   import useLocale from '../composables/useLocale'
 
   const props = withDefaults(
     defineProps<{
       text: K
-      values?: V
+      context?: V
       tag?: string
     }>(),
     { tag: 'span' },
   )
 
-  type SlotProps = typeof props.values & { decline: FormatContext<any>['decline'] }
+  type SlotProps = typeof props.context & { inflect: ResolveContext<any>['inflect'] }
   type Slots = { [key in keyof V]: (slotProps: SlotProps) => any }
   defineSlots<Slots>()
 
   const slots = useSlots()
 
-  const values = props.values || {}
+  const context = props.context || {}
   const locale = useLocale()
 
   const segments = computed(() => {
     const localeValue = locale.value
-    const { localizedMessage, formatOptions } = prepareFormat(props.text as string, localeValue)
-    const segments = formatToSegments(localizedMessage, values, formatOptions)
+    const localizedMessage = localizeMessage(props.text as string, localeValue)
+    const segments = resolveToSegments(localizedMessage, context, {
+      locale: localeValue,
+    })
 
     return segments
   })
 
   onMounted(() => {
     if (import.meta.env.DEV)
-      collect(props.text as any, props.values as Values)
+      collect(props.text as any, props.context as Context)
   })
 </script>
 
@@ -44,7 +46,7 @@
         {{ segment.part }}
       </template>
       <template v-else>
-        <slot :name="(segment.key!)" v-bind="(segment.values as V)" :decline="(segment.decline)" />
+        <slot :name="(segment.key!)" v-bind="(segment.context as V)" :inflect="(segment.inflect)" />
       </template>
     </template>
   </component>